Allinone: test cleanup
* Unit tests should not be here, so moved a bunch of those next to the code they test
* Test names should make sense and be unique. Specifically, unified on
  <QuestionName>Test for e2e tests and <ClassName>Test for unit tests.
  <ClassName> is typically FooQuestion, FooAnswer, where as
  <QuestionName> is Foo.
* Add javadocs.
