+++ projects/allinone/src/test/java/org/batfish/question/reachfilter/ReachFilterTest.java
@@ -66,7 +66,8 @@ import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-public final class ReachFilterAnswererTest {
+/** End-to-end tests of {@link ReachFilterQuestion}. */
+public final class ReachFilterTest {
   private static final String IFACE1 = "iface1";
 
   private static final String IFACE2 = "iface2";
