+++ projects/question/src/test/java/org/batfish/question/traceroute/TracerouteAnswererTest.java
@@ -10,6 +10,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multiset;
+import java.util.List;
 import java.util.Set;
 import org.batfish.datamodel.Flow;
 import org.batfish.datamodel.FlowDisposition;
@@ -17,12 +18,63 @@ import org.batfish.datamodel.FlowHistory;
 import org.batfish.datamodel.FlowHistory.FlowHistoryInfo;
 import org.batfish.datamodel.FlowTrace;
 import org.batfish.datamodel.Ip;
+import org.batfish.datamodel.answers.Schema;
 import org.batfish.datamodel.pojo.Environment;
+import org.batfish.datamodel.table.ColumnMetadata;
 import org.batfish.datamodel.table.Row;
 import org.batfish.datamodel.table.TableDiff;
+import org.batfish.datamodel.table.TableMetadata;
 import org.junit.Test;
 
 public class TracerouteAnswererTest {
+
+  @Test
+  public void testCreateMetadata() {
+    TableMetadata tableMetadata = TracerouteAnswerer.createMetadata(false);
+    List<String> columnNames =
+        tableMetadata
+            .getColumnMetadata()
+            .stream()
+            .map(ColumnMetadata::getName)
+            .collect(ImmutableList.toImmutableList());
+    List<Schema> columnSchemas =
+        tableMetadata
+            .getColumnMetadata()
+            .stream()
+            .map(ColumnMetadata::getSchema)
+            .collect(ImmutableList.toImmutableList());
+
+    assertThat(columnNames, equalTo(ImmutableList.of(COL_FLOW, COL_TRACES)));
+    assertThat(
+        columnSchemas, equalTo(ImmutableList.of(Schema.FLOW, Schema.set(Schema.FLOW_TRACE))));
+
+    TableMetadata diffTableMetadata = TracerouteAnswerer.createMetadata(true);
+    List<String> diffColumnNames =
+        diffTableMetadata
+            .getColumnMetadata()
+            .stream()
+            .map(ColumnMetadata::getName)
+            .collect(ImmutableList.toImmutableList());
+    List<Schema> diffColumnSchemas =
+        diffTableMetadata
+            .getColumnMetadata()
+            .stream()
+            .map(ColumnMetadata::getSchema)
+            .collect(ImmutableList.toImmutableList());
+
+    assertThat(
+        diffColumnNames,
+        equalTo(
+            ImmutableList.of(
+                COL_FLOW,
+                TableDiff.baseColumnName(COL_TRACES),
+                TableDiff.deltaColumnName(COL_TRACES))));
+    assertThat(
+        diffColumnSchemas,
+        equalTo(
+            ImmutableList.of(
+                Schema.FLOW, Schema.set(Schema.FLOW_TRACE), Schema.set(Schema.FLOW_TRACE))));
+  }
 
   @Test
   public void testDiffFlowHistoryToRows() {
